+++ BE/src/main/java/roomescape/infrastructure/ReservationJdbcTemplateRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservations.infrastructure;
+package roomescape.infrastructure;
 
 import java.util.List;
 import java.util.Map;
@@ -7,9 +7,9 @@ import javax.sql.DataSource;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.reservations.entity.Reservation;
-import roomescape.reservations.entity.ReservationRepository;
-import roomescape.reservations.entity.ReservationTime;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationRepository;
+import roomescape.entity.ReservationTime;
 
 @Repository
 public class ReservationJdbcTemplateRepository implements ReservationRepository {
