+++ BE/src/main/java/roomescape/infrastructure/ReservationTimeJdbcTemplateRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservations.infrastructure;
+package roomescape.infrastructure;
 
 import java.util.List;
 import java.util.Map;
@@ -7,8 +7,8 @@ import javax.sql.DataSource;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.reservations.entity.ReservationTime;
-import roomescape.reservations.entity.ReservationTimeRepository;
+import roomescape.entity.ReservationTime;
+import roomescape.entity.ReservationTimeRepository;
 
 @Repository
 public class ReservationTimeJdbcTemplateRepository implements ReservationTimeRepository {
