+++ BE/src/test/java/roomescape/FakeReservationTimeRepository.java
@@ -1,11 +1,11 @@
-package roomescape.domain.reservations;
+package roomescape;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import roomescape.reservations.entity.ReservationTime;
-import roomescape.reservations.entity.ReservationTimeRepository;
+import roomescape.entity.ReservationTime;
+import roomescape.entity.ReservationTimeRepository;
 
 public class FakeReservationTimeRepository implements ReservationTimeRepository {
